Make kits work correctly. For some reason, yaml doesn't like uppercase configuration section keys.
