+++ Essentials/kits.yml
@@ -1,21 +1,6 @@
 # EssentialsX new Kit configuration.
 # If you don't have any kits defined in this file, the plugin will try to copy them from the config.yml
 
-
-
-# When this option is enabled, one-time use kits (ie. delay < 0) will be
-# removed from the /kit list when a player can no longer use it
-skip-used-one-time-kits-from-kit-list: false
-
-# Determines the functionality of the /createkit command.
-# If this is true, /createkit will give the user a link with the kit code.
-# If this is false, /createkit will add the kit to this config file directly.
-#
-# WARNING: If this is false, the comments in this file WILL be removed and it won't look the same as it does now.
-pastebin-createkit: true
-
-
-
 # Note: All items MUST be followed by a quantity!
 # All kit names should be lower case, and will be treated as lower in permissions/costs.
 # Syntax: - itemID[:DataValue/Durability] Amount [Enchantment:Level].. [itemmeta:value]...
@@ -23,31 +8,32 @@ pastebin-createkit: true
 # 'delay' refers to the cooldown between how often you can use each kit, measured in seconds.
 # Set delay to -1 for a one time kit.
 # For more information, visit http://wiki.ess3.net/wiki/Kits
-tools:
-  delay: 10
-  items:
-    - 272 1
-    - 273 1
-    - 274 1
-    - 275 1
-dtools:
-  delay: 600
-  items:
-    - 278 1 efficiency:1 durability:1 fortune:1 name:&4Gigadrill lore:The_drill_that_&npierces|the_heavens
-    - 277 1 digspeed:3 name:Dwarf lore:Diggy|Diggy|Hole
-    - 298 1 color:255,255,255 name:Top_Hat lore:Good_day,_Good_day
-    - 279:780 1
-notch:
-  delay: 6000
-  items:
-    - 397:3 1 player:Notch
-color:
-  delay: 6000
-  items:
-    - 387 1 title:&4Book_&9o_&6Colors author:KHobbits lore:Ingame_color_codes book:Colors
-firework:
-  delay: 6000
-  items:
-    - 401 1 name:Angry_Creeper color:red fade:green type:creeper power:1
-    - 401 1 name:Starry_Night color:yellow,orange fade:blue type:star effect:trail,twinkle power:1
-    - 401 2 name:Solar_Wind color:yellow,orange fade:red shape:large effect:twinkle color:yellow,orange fade:red shape:ball effect:trail color:red,purple fade:pink shape:star effect:trail power:1
+kits:
+  tools:
+    delay: 10
+    items:
+      - 272 1
+      - 273 1
+      - 274 1
+      - 275 1
+  dtools:
+    delay: 600
+    items:
+      - 278 1 efficiency:1 durability:1 fortune:1 name:&4Gigadrill lore:The_drill_that_&npierces|the_heavens
+      - 277 1 digspeed:3 name:Dwarf lore:Diggy|Diggy|Hole
+      - 298 1 color:255,255,255 name:Top_Hat lore:Good_day,_Good_day
+      - 279:780 1
+  notch:
+    delay: 6000
+    items:
+      - 397:3 1 player:Notch
+  color:
+    delay: 6000
+    items:
+      - 387 1 title:&4Book_&9o_&6Colors author:KHobbits lore:Ingame_color_codes book:Colors
+  firework:
+    delay: 6000
+    items:
+      - 401 1 name:Angry_Creeper color:red fade:green type:creeper power:1
+      - 401 1 name:Starry_Night color:yellow,orange fade:blue type:star effect:trail,twinkle power:1
+      - 401 2 name:Solar_Wind color:yellow,orange fade:red shape:large effect:twinkle color:yellow,orange fade:red shape:ball effect:trail color:red,purple fade:pink shape:star effect:trail power:1
