--- Essentials/src/com/earth2me/essentials/IKitSettings.java
@@ -1,8 +1,0 @@
-package com.earth2me.essentials;
-
-public interface IKitSettings {
-
-    boolean isSkippingUsedOneTimeKitsFromKitList();
-
-    boolean isPastebinCreateKit();
-}
